Add a small webserver to g-d-c
This patch adds a small webserver based on the bottle framework
to gerrit-dash-creator.

If you run the script with the "--webserver" flag then it will
serve up a list of URLS generated from the dashboards directory at
http://localhost:8080/

If you add/remove/update a dashboard then you can just refresh
the page in your browser and the changes will show up in the list.

There is also a JSON endpoint available at
http://localhost:8080/json/

Change-Id: Ic41217a5214e33dfae1fd4590e367cb840b04126