Add a 'possible abandoned changes' query to swift
This change adds a new section called 'Possible abandoned changes'
to the swift dashboard. This query looks for open changes that have
a -1, that isn't in WIP and hasn't been touched in 4 weeks:

  status:open age:4week (label:Verified<=-1,jenkins OR
     label:Code-Review<=-1) NOT label:Workflow<=-1

This will be used to keep track of patches that look to be abandoned
so that they can be.

Change-Id: I754e790457fbf9f59d2974511b7aa6f65f2ff06f