Add py34 env to tox.ini
Python 3.4 is the default in Ubuntu Trusty and the OpenStack test
infrastructure is going to drop Python 3.3 testing in favor of that.
In an effort to make 3.4 test results more visible to developers who
run the default tox env list and happen to have 3.4 installed, add a
py34 env for it.

Change-Id: Id778c0f1513a1db7ad351186547b792b3b116c6e