Automatically generate a dash based on "In Progress" bugs
This adds a script to get the "In Progress" bugs for a
particular project and builds prints a URL for the user.

 $ ./gerrit-bug-dash --project heat juno-rc1

Some notes:
 - it uses launchpadlib to talk to launchpad
 - The presentation of the dashboard could be improved but
   this is just a starting point.
 - The caching could be better.

What I wanted was to know was "what do I need to review for juno-rc1"
(i.e. what are the bugs that are targeted for juno-rc1 that are in
progress and have reviews up).

This scratches my itch, but will happily do some work on it if others
want it.

Change-Id: I4a97d59631ac9cd344206c6cc48164d6a0d7e57c