update nova-specs dashboard
The specs dashboard was hiding important things. For instance -1 on a
spec isn't actually a reason to not review it, because we should be
providing all the feedback at once, not making there be 20 iterations
of a spec.

We should also show things that are wedged by Jenkins, because that
might be a merge conflict or some other pretty innocuous change that
we should just fix.

The "my specs" section was also kind of dubious, because you aren't
going to review your own specs.

Change-Id: Idef5de59f2c207d95e453be1b6fc8f3778254c32