Add ability to dynamically generate the foreach projects
Add support for being able to use {{projects}} in the foreach line.  If
that is specified and 'project' is set to the project name (e.g. Ironic)
then the project governance file will be downloaded and used to
determine all the project repositories for the project.

Also update dashboards/ironic.dash to use this feature.

Change-Id: I8b9e5e416818a8b4599d4964922fe008796124d7