Fixed reviewedby:self logic
""You are a reviewer, but haven't voted in the current revision" section
was showing patches that already had votes from self. Using
reviewedby:self doesn't help either because it matches against patches
that you voted for but someone important (like the author) happened to
comment after your vote. Hence the usage of Code-Review matching rule.

This was noticed for neutron dashboard but I fixed it in a bunch of
other places.

Change-Id: I0bd2a646df1e8f268472efcd740a9d3d51aea962