Add Cinder review dashboard
Some basic review categories for Cinder. There is actually an issue right
now with the way this (and other dashboards) are querying bug fixes that
appears to have been introduced with the last gerrit upgrade. Even though
these do not return and matches right now, including them here as a
breadcrumb to investigate later, or in case someone else does and fixes
it for all the dashboards.

Change-Id: If8ecefb1e5841decdcb1bcb046cb054bd9ed9175